(fix) Norminette OK

+++ src/exec/launch_program.c
@@ -6,7 +6,7 @@
 /*   By: erpascua <erpascua@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/11 18:37:31 by erpascua          #+#    #+#             */
-/*   Updated: 2025/08/13 19:30:20 by erpascua         ###   ########.fr       */
+/*   Updated: 2025/08/14 18:55:27 by erpascua         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,22 +18,17 @@ void	update_history(t_msh *msh, char *entry)
 	append_history(1, msh->history);
 }
 
-int	launch_program(t_msh *msh)
+void	repl(t_msh *msh, int tmp_fd)
 {
 	char	*entry;
-	int		tmp_fd;
 
-	read_history(msh->history);
-	tmp_fd = open("tmp_fd", O_WRONLY | O_CREAT | O_TRUNC, 0644);
-	if (tmp_fd < 0)
-		return (perror("open"), 1);
 	while (1)
 	{
-		if (isatty(0))
+		if (isatty(STDIN_FILENO))
 			entry = readline("\033[1;92mMinishell > \033[0m");
 		else
 			entry = get_next_line(0);
-		if (!entry)
+		if (!entry && is_eof())
 			break ;
 		else
 		{
@@ -43,8 +38,19 @@ int	launch_program(t_msh *msh)
 		write(tmp_fd, entry, ft_strlen(entry));
 		free(entry);
 	}
+}
+
+int	launch_program(t_msh *msh)
+{
+	int		tmp_fd;
+
+	read_history(msh->history);
+	tmp_fd = open("tmp_fd", O_WRONLY | O_CREAT | O_TRUNC, 0644);
+	if (tmp_fd < 0)
+		return (perror("open"), 1);
+	repl(msh, tmp_fd);
 	close(tmp_fd);
 	free(msh->history);
 	unlink("tmp_fd");
-	return (0);
+	return (msh->last_status);
 }
